Add manual controls to external-HR workouts
Share the touch-friendly trainer target control between free rides and workouts. Show Power and Resistance adjustment during cycling trainer workouts with heart-rate targets when the configured HR sensor is separate from the trainer.

Add helpers and regression coverage for detecting heart-rate workouts and distinguishing external trainer HR sensors.
