+++ src/fitness_tracker/workouts.py
@@ -8,6 +8,28 @@ from pathlib import Path
 AUTO_SUBDIRS = ("intervals_icu",)
 
 
+def has_heart_rate_targets(steps) -> bool:
+    """Return whether any workout step contains a supported heart-rate target."""
+    fields = (
+        "heart_rate_bpm",
+        "heart_rate_percent_max",
+        "heart_rate_percent_lthr",
+        "heart_rate_zone",
+    )
+    return any(any(getattr(step, field, None) is not None for field in fields) for step in steps)
+
+
+def has_external_trainer_hr_sensor(recorder) -> bool:
+    """Return whether the configured trainer HR source is a separate device."""
+    if not recorder:
+        return False
+    if recorder.hr_address and recorder.trainer_address:
+        return recorder.hr_address != recorder.trainer_address
+    if recorder.hr_name and recorder.trainer_name:
+        return recorder.hr_name != recorder.trainer_name
+    return False
+
+
 def apply_target_bias(
     values: tuple[float, float, float] | None,
     percent: int,
